requirements.txt
pillow
openai
requests
python-dotenv